briteleader-new/requirements.txt
asgi-redis==1.4.3
asgiref==3.2.3
amqp==2.3.2
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
billiard==3.5.0.4
bleach==3.1.0
boto3==1.9.33
botocore==1.12.33
CacheControl==0.12.5
cachetools==3.1.0
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
cssselect==1.0.3
cssutils==1.0.2
daphne==2.4.1
decorator==4.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
dj-stripe==2.2.1
Django==2.2.8
django-allauth==0.36.0
django-babel==0.6.2
django-bootstrap4==0.0.7
django-cors==0.1
django-cors-headers==2.2.0
django-easy-pdf==0.1.1
django-extensions==2.1.9
django-filter==2.2.0
django-filters==0.2.1
django-heroku==0.3.1
django-ipware==2.1.0
django-mathfilters==0.4.0
django-modeltranslation==0.13.1
django-money==0.14.4
django-prices==1.0.4
django-prices-openexchangerates==1.0.1
django-rest==0.0.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-seed==0.1.9
django-storages==1.7.1
django-taggit==0.23.0
django-uuidfield==0.5.0
django-webpack-loader==0.6.0
django-webpush==0.2.4
djangorestframework==3.8.2
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
docutils==0.14
drf-yasg==1.15.0
entrypoints==0.3
Faker==0.8.15
fcm-django==0.2.21
firebase-admin==2.16.0
flex==6.13.2
future==0.16.0
geocoder==1.38.1
geographiclib==1.49
geopy==1.17.0
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
googlemaps==3.0.2
grpcio==1.19.0
gunicorn==19.9.0
http-ece==1.0.5
idna==2.6
inflection==0.3.1
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonpointer==1.14
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.2.1
lxml==4.2.5
MarkupSafe==1.0
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
oauthlib==2.1.0
openapi-codec==1.3.2
pandocfilters==1.4.2
parso==0.3.1
paypalrestsdk==1.13.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.1.0
premailer==3.2.0
prices==1.0.0
prometheus-client==0.7.0
prompt-toolkit==2.0.4
protobuf==3.7.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pudb==2018.1
py-moneyed==0.8.0
py-vapid==1.4.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.18
pyfcm==1.4.5
Pygments
PyJWT==1.6.4
PyMySQL==0.9.3
pyOpenSSL==18.0.0
PySocks==1.7.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
pywebpush==1.6.0
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.5.1
ratelim==0.1.6
redis==2.10.6
requests==2.21.0
requests-oauthlib==0.8.0
requests-toolbelt==0.9.1
rfc3987==1.3.7
rsa==4.0
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
twilio==6.26.3
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
urwid==2.0.1
validate-email==1.3
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1
widgetsnbextension==3.4.2
xhtml2pdf==0.2.3
zipp==0.6.0
zope.interface==4.7.1
